+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * Aligning development and deployment Dockerfiles [#38](https://gitlab.data.bas.ac.uk/MAGIC/web-map-inventory/issues/38)
 * Improving deployment container [#38](https://gitlab.data.bas.ac.uk/MAGIC/web-map-inventory/issues/38) [#51](https://gitlab.data.bas.ac.uk/MAGIC/web-map-inventory/issues/51)
+* Improved Podman wrapper script [#48](https://gitlab.data.bas.ac.uk/MAGIC/web-map-inventory/issues/48)
 
 ## [0.2.1] - 2020-01-02
 
@@ -55,8 +56,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * Switching from Pip (for dependencies) and setuptools (for packaging) to Poetry [#27](https://gitlab.data.bas.ac.uk/MAGIC/web-map-inventory/issues/27) 
 * Switching from Flake8 to Black for PEP8 compliance/formatting [#27](https://gitlab.data.bas.ac.uk/MAGIC/web-map-inventory/issues/27) 
-* Docker image changed to run as a non-privileged user 
-  [#11](https://gitlab.data.bas.ac.uk/MAGIC/web-map-inventory/issues/11)
+* Docker image changed to run as a non-privileged user [#11](https://gitlab.data.bas.ac.uk/MAGIC/web-map-inventory/issues/11)
 
 ## [0.1.0] - 2019-11-25